Refactor foodData step components
