+++ components/foodData/VegetableProportion.js
@@ -1,6 +1,3 @@
-import styled from 'styled-components'
-import { Field } from 'formik'
-
 import allIcon from '../../public/icons/quantities/regular/all.svg'
 import allIconSelected from '../../public/icons/quantities/selected/all-selected.svg'
 import halfIcon from '../../public/icons/quantities/regular/half.svg'
@@ -11,82 +8,41 @@ import quarterIcon from '../../public/icons/quantities/regular/quarter.svg'
 import quarterIconSelected from '../../public/icons/quantities/selected/quarter-selected.svg'
 import vegIcon from '../../public/icons/categories/regular/vegan.svg'
 
+import {
+  Title,
+  CardBackground,
+  RadioTile,
+  TileContainer,
+  FruitVegTile,
+} from './shared'
+
 import * as Steps from '.'
-
-const RadioButton = ({
-  field: { name, value, onChange, onBlur },
-  id,
-  label,
-  ...props
-}) => {
-  return (
-    <input
-      name={name}
-      id={id}
-      type="radio"
-      value={id} // could be something else for output?
-      checked={id === value}
-      onChange={onChange}
-      onBlur={onBlur}
-      {...props}
-    />
-  )
-}
-
-const ProportionButton = ({ name, icon, iconSelected, label, id }) => {
-  const text = label.charAt(0).toUpperCase() + label.slice(1)
-  return (
-    <RadioContainer
-      htmlFor={label}
-      className="parent"
-      icon={icon}
-      iconSelected={iconSelected}
-    >
-      <Field name={name} label={label} component={RadioButton} id={id} />
-      <div className="background">
-        <span className="checkmark" />
-        <span>{text}</span>
-      </div>
-    </RadioContainer>
-  )
-}
-
-const VegetableTile = styled.div`
-  width: 190px;
-  height: 6rem;
-  border: 1px solid black;
-  border-radius: 1rem;
-
-  .checkmark {
-    background-image: url(${vegIcon});
-    height: 4rem;
-    background-repeat: no-repeat;
-    background-position: center;
-  }
-`
 
 const VegetableProportion = () => {
   return (
     <CardBackground>
       <Title>You tagged Vegetables</Title>
-      <VegetableTile className="flex flex-col justify-center space-between mx-auto mb-5">
+      <FruitVegTile
+        className="flex flex-col justify-center space-between mx-auto mb-5"
+        icon={vegIcon}
+      >
         <div className="checkmark" />
         <p className="w-full text-center">Vegetables</p>
-      </VegetableTile>
+      </FruitVegTile>
 
       <Title className="w-11/12 mx-auto">
         Roughly, how much of the plate is vegetables?
       </Title>
 
-      <IconContainer>
-        <ProportionButton
+      <TileContainer>
+        <RadioTile
           name="proportionVeg"
           label="quarter"
           id="quarter"
           icon={quarterIcon}
           iconSelected={quarterIconSelected}
         />
-        <ProportionButton
+        <RadioTile
           name="proportionVeg"
           label="half"
           id="half"
@@ -94,7 +50,7 @@ const VegetableProportion = () => {
           iconSelected={halfIconSelected}
         />
 
-        <ProportionButton
+        <RadioTile
           name="proportionVeg"
           label="mostly"
           id="mostly"
@@ -102,87 +58,17 @@ const VegetableProportion = () => {
           iconSelected={mostlyIconSelected}
         />
 
-        <ProportionButton
+        <RadioTile
           name="proportionVeg"
           label="all"
           id="all"
           icon={allIcon}
           iconSelected={allIconSelected}
         />
-      </IconContainer>
+      </TileContainer>
     </CardBackground>
   )
 }
-
-const RadioContainer = styled.label.attrs({
-  className: 'block relative cursor-pointer select-none',
-})`
-  width: 100%;
-  /* Hide the browser's default checkbox */
-  input {
-    position: absolute;
-    opacity: 0;
-    cursor: pointer;
-    height: 0;
-    width: 0;
-  }
-
-  div {
-    width: 100%;
-    height: 6rem;
-    border: 1px solid black;
-    border-radius: 1rem;
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: center;
-    text-align: center;
-  }
-
-  /* Create a custom checkbox */
-  .checkmark {
-    width: 100%;
-    height: 4rem;
-    background-image: url(${props => props.icon});
-    background-repeat: no-repeat;
-    background-position: center;
-    display: block;
-  }
-  /* When the checkbox is checked, add a blue background */
-  input:checked ~ .background > .checkmark {
-    background-image: url(${props => props.iconSelected});
-  }
-
-  input:checked ~ .background {
-    background-color: ${cssTheme('colors.navy')};
-    color: white;
-  }
-
-  input:checked .parent {
-    background-color: blue;
-  }
-`
-
-const CardBackground = styled.section.attrs({
-  className: 'z-10 w-screen bg-white',
-})`
-  border-top-left-radius: 4em;
-  border-top-right-radius: 4em;
-`
-
-const Title = styled.h1.attrs({
-  className: 'font-xxl text-center mb-5 mt-5',
-})``
-
-const IconContainer = styled.section.attrs({
-  className: '',
-})`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: 1rem;
-  justify-items: center;
-  max-width: 90%;
-  margin: 0 auto;
-`
 
 VegetableProportion.componentName = Steps.VegetableProportion
